Created Fakes namespace and put fake classes in their own files #15.1
