+++ src/Minicloner/Minicloner.Tests/Fakes/ClassWithPublicAutomaticallyImplementedProperty.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Minicloner.Tests.Fakes
+{
+    public class ClassWithPublicAutomaticallyImplementedProperty
+    {
+        public Int32 PublicInt32Property { get; set; }
+    }
+}
