+++ src/Minicloner/Minicloner.Tests/Fakes/ClassWithPublicValueTypeField.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Minicloner.Tests.Fakes
+{
+    public class ClassWithPublicValueTypeField
+    {
+        public Int32 PublicInt32Field;
+    }
+}
