+++ src/Minicloner/Minicloner.Tests/Fakes/ClassWithReferenceTypeProperty.cs
@@ -1,0 +1,7 @@
+namespace Minicloner.Tests.Fakes
+{
+    public class ClassWithReferenceTypeProperty
+    {
+        public ClassWithPublicAutomaticallyImplementedProperty ReferenceTypeProperty { get; set; }
+    }
+}
