+++ src/Minicloner/Minicloner.Tests/Fakes/ClassWithTwoLevelsOfReferenceTypeProperties.cs
@@ -1,0 +1,8 @@
+namespace Minicloner.Tests.Fakes
+{
+    public class ClassWithTwoLevelsOfReferenceTypeProperties
+    {
+        public ClassWithPublicAutomaticallyImplementedProperty ReferenceTypeProperty { get; set; }
+        public ClassWithReferenceTypeProperty TwoLeveledReferenceTypeProperty { get; set; }
+    }
+}
