+++ src/Minicloner/Minicloner.Tests/Fakes/ClassWithoutParameterlessConstructor.cs
@@ -1,0 +1,9 @@
+namespace Minicloner.Tests.Fakes
+{
+    public class ClassWithoutParameterlessConstructor
+    {
+        public ClassWithoutParameterlessConstructor(object dummy)
+        {
+        }
+    }
+}
