+++ src/Minicloner/Minicloner.Tests/Fakes/EmptyClass.cs
@@ -1,0 +1,6 @@
+namespace Minicloner.Tests.Fakes
+{
+    public class EmptyClass
+    {
+    }
+}
