+++ src/Minicloner/Minicloner.Tests/Minicloner.Tests.csproj
@@ -51,6 +51,15 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Fakes\ClassWithoutParameterlessConstructor.cs" />
+    <Compile Include="Fakes\ClassWithPrivateGetter.cs" />
+    <Compile Include="Fakes\ClassWithPrivateSetter.cs" />
+    <Compile Include="Fakes\ClassWithPrivateValueTypeField.cs" />
+    <Compile Include="Fakes\ClassWithPublicAutomaticallyImplementedProperty.cs" />
+    <Compile Include="Fakes\ClassWithPublicValueTypeField.cs" />
+    <Compile Include="Fakes\ClassWithReferenceTypeProperty.cs" />
+    <Compile Include="Fakes\ClassWithTwoLevelsOfReferenceTypeProperties.cs" />
+    <Compile Include="Fakes\EmptyClass.cs" />
     <Compile Include="ClonerArrayTests.cs" />
     <Compile Include="ClonerClassesTests.cs" />
     <Compile Include="ClonerDateTimeTests.cs" />
@@ -74,6 +83,7 @@
       <Name>Minicloner</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
